Tests passing
When I visit /d20 the page loads correctly but the response printed out in the console has a bunch of garbage data following the random number. None of the other pages have this problem and it only started after implementing the cache. Seems like there might be a malloc issue in alloc_entry but I can't figure it out for the moment.